Fixed some stuff and added file with functions
